Rotation of Point
Script to rotate an equiretangular point about the specified axis. The
user determines the angle of rotation and the axis to rotate the point
about. This point represents a single pixel within an image. It has been
tested for rotations of various points at various angles about all the
axes.